requirements.txt
alembic==1.11.1
amqp==5.1.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.2
asyncpg==0.28.0
attrs==23.1.0
bcrypt==4.0.1
billiard==4.1.0
black==23.7.0
celery==5.3.4
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.4
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coverage==7.3.0
cryptography==41.0.2
distro==1.8.0
dnspython==2.3.0
docker==6.1.3
dockerpty==0.4.1
docopt==0.6.2
email-validator==2.0.0.post2
exceptiongroup==1.1.2
execnet==2.0.2
fastapi==0.100.0
fastapi-pagination==0.12.6
fastapi-users==12.1.0
fastapi-users-db-sqlalchemy==6.0.0
flake8==6.1.0
flower==2.0.1
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.3
httptools==0.6.0
httpx==0.24.1
humanize==4.8.0
idna==3.4
iniconfig==2.0.0
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==3.2.0
kombu==5.3.2
makefun==1.15.1
Mako==1.2.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mccabe==0.7.0
mdurl==0.1.2
mypy==1.4.1
mypy-extensions==1.0.0
orjson==3.9.2
packaging==23.1
paramiko==3.2.0
passlib==1.7.4
pathspec==0.11.2
platformdirs==3.10.0
pluggy==1.2.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
py-dotenv==0.1
pycodestyle==2.11.1
pycparser==2.21
pydantic==2.0.2
pydantic-extra-types==2.0.0
pydantic-settings==2.0.1
pydantic_core==2.1.2
pyflakes==3.1.0
Pygments==2.16.1
PyJWT==2.7.0
PyNaCl==1.5.0
pyrsistent==0.19.3
pytest==7.4.0
pytest-asyncio==0.21.0
pytest-cov==4.1.0
pytest-picked==0.5.0
pytest-sugar==0.9.7
pytest-xdist==3.3.1
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==0.21.1
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
redis==4.6.0
requests==2.31.0
rich==13.5.2
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.18
starlette==0.27.0
termcolor==2.3.0
texttable==1.6.7
tomli==2.0.1
tornado==6.3.3
typing_extensions==4.7.1
tzdata==2023.3
ujson==5.8.0
urllib3==2.0.3
uvicorn==0.22.0
uvloop==0.17.0
vine==5.0.0
watchfiles==0.19.0
wcwidth==0.2.6
websocket-client==0.59.0
websockets==11.0.3
